Rigidbody

+++ ACT-Project_DX11/Engine/TimeManager.h
@@ -20,6 +20,6 @@ private:
 private:
 	uint32	_frameCount = 0;
 	float	_frameTime = 0.f;
-	uint32	_fps = 0;
+	float	_fps = 0;
 };
